+++ src/components/FormInputGroups/ContactGroup.tsx
@@ -3,8 +3,8 @@ import { Box, Button, Divider } from "@mui/material";
 import { useContext } from "react";
 import { useFieldArray } from "react-hook-form";
 import { TextFieldElement } from "react-hook-form-mui";
-import ReadOnlyContext from "../../services/ReadOnlyContext";
-import { ContactIndividual } from "../../utilities/globalTypes";
+import { ContactIndividual } from "../../functions/globalTypes";
+import ReadOnlyContext from "../contexts/ReadOnlyContext";
 import PhoneNumber from "../inputs/PhoneNumber";
 import TextEditor from "../inputs/TextEditor";
 import { TitledGroup } from "../layouts/TitledGroup";
