+++ src/components/FormInputGroups/ProgramGroup.tsx
@@ -3,8 +3,8 @@ import { Box, Button, Divider, Typography } from "@mui/material";
 import { useContext } from "react";
 import { useFieldArray, useWatch } from "react-hook-form";
 import { TextFieldElement } from "react-hook-form-mui";
-import ReadOnlyContext from "../../services/ReadOnlyContext";
-import { EventComponent } from "../../utilities/globalTypes";
+import { EventComponent } from "../../functions/globalTypes";
+import ReadOnlyContext from "../contexts/ReadOnlyContext";
 import ControlledCheckbox from "../inputs/ControlledCheckbox";
 import ControlledDate from "../inputs/ControlledDate";
 import ControlledTime from "../inputs/ControlledTime";
