+++ src/components/inputs/ControlledTime.tsx
@@ -2,8 +2,8 @@ import { AdapterDateFns } from "@mui/x-date-pickers/AdapterDateFns";
 import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
 import { TimePicker } from "@mui/x-date-pickers/TimePicker";
 import { Controller, FieldPath, useFormContext } from "react-hook-form";
-import filterDateOrNumberToDate from "../../utilities/filterDateOrNumberToDate";
-import { AllEventGroups } from "../../utilities/globalTypes";
+import filterDateOrNumberToDate from "../../functions/filterDateOrNumberToDate";
+import { AllEventGroups } from "../../functions/globalTypes";
 
 interface Props {
   dataLabel: FieldPath<AllEventGroups>;
