+++ src/components/inputs/TimeRangePicker.tsx
@@ -1,6 +1,6 @@
 import { Box } from "@mui/material";
 import { FieldPath, useFormContext } from "react-hook-form";
-import { AllEventGroups } from "../../utilities/globalTypes";
+import { AllEventGroups } from "../../functions/globalTypes";
 import ControlledTime from "./ControlledTime";
 
 export default function TimeRangePicker({
