+++ src/functions/globalTypes.ts
